+++ src/ringhopper-primitives/src/primitive/plane.rs
@@ -8,14 +8,14 @@ pub trait Plane {
     type VectorType: Vector;
 
     fn vector(&self) -> Self::VectorType;
-    fn d(&self) -> f32;
+    fn d(&self) -> f64;
 }
 
 #[derive(Clone, Copy, Default, Debug, PartialEq)]
 #[repr(C)]
 pub struct Plane2D {
     pub vector: Vector2D,
-    pub d: f32
+    pub d: f64
 }
 
 impl Display for Plane2D {
@@ -29,7 +29,7 @@ impl Plane for Plane2D {
     fn vector(&self) -> Vector2D {
         self.vector
     }
-    fn d(&self) -> f32 {
+    fn d(&self) -> f64 {
         self.d
     }
 }
@@ -40,7 +40,7 @@ impl SimpleTagData for Plane2D {
     }
     fn read<B: ByteOrder>(data: &[u8], at: usize, struct_end: usize) -> RinghopperResult<Self> {
         let vector = Vector2D::read::<B>(data, at, struct_end)?;
-        let d = f32::read::<B>(data, at + 0x8, struct_end)?;
+        let d = f64::read::<B>(data, at + 0x8, struct_end)?;
         Ok(Self {
             vector, d
         })
@@ -62,7 +62,7 @@ impl SimplePrimitive for Plane2D {
 #[repr(C)]
 pub struct Plane3D {
     pub vector: Vector3D,
-    pub d: f32
+    pub d: f64
 }
 
 impl Plane for Plane3D {
@@ -70,7 +70,7 @@ impl Plane for Plane3D {
     fn vector(&self) -> Vector3D {
         self.vector
     }
-    fn d(&self) -> f32 {
+    fn d(&self) -> f64 {
         self.d
     }
 }
@@ -81,7 +81,7 @@ impl SimpleTagData for Plane3D {
     }
     fn read<B: ByteOrder>(data: &[u8], at: usize, struct_end: usize) -> RinghopperResult<Self> {
         let vector = Vector3D::read::<B>(data, at, struct_end)?;
-        let d = f32::read::<B>(data, at + 0xC, struct_end)?;
+        let d = f64::read::<B>(data, at + 0xC, struct_end)?;
         Ok(Self {
             vector, d
         })
